[Refactoring] : 유닛 시스템 리팩토링
