--- Assets/Scripts/Player/PlayerAction.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 39e87734464af4d48a0fab203a812326
