+++ Assets/Scripts/Player/PlayerController.cs
@@ -1,12 +1,16 @@
 using System.Collections;
 using Unity.Cinemachine;
+using Unity.Netcode;
 using UnityEngine;
 
 public class PlayerController : UnitController
 {
+    [SerializeField] private Animator _modelAnimator;
+    [SerializeField] private LayerMask _groundMask;
+
     [SerializeField] private Transform _respawnAnchor;//나중에 할당으로 수정해야 함
 
-    private PlayerAction _playerAction;
+
     /// <summary>
     /// 레벨에 따른 부활시간 RESAPWN_TIME[현재레벨], 단위는 초
     /// </summary>
@@ -14,78 +18,43 @@ public class PlayerController : UnitController
 
     public bool IsAttackButtonDown { get; private set; }
 
-    private new void Awake()
+    const float MOVE_STOPPING_DISTANCE = 3.0f;
+
+    protected override void Awake()
     {
         base.Awake();
-        _playerAction = GetComponent<PlayerAction>();
     }
 
     private new void Start()
     {
         base.Start();
 
-        HeroHpBarUI heroHpBarUI = _unitHPBarUI as HeroHpBarUI;
+        PlayerInputManager.Instance.OnLeftClickEvent.AddListener(OnLeftMouseDown);
+        PlayerInputManager.Instance.OnRightClickEvent.AddListener(OnRightMouseDown);
+        PlayerInputManager.Instance.OnAttackButtonEvent.AddListener(OnAttackButtonDown);
+    }
 
-        Logger.Info($"영웅 이름 : {GetHeroName()}");
-        Logger.Info($"현재 레벨 : {GetLevel()}");
+    public void Init(UnitTeamType team, ulong clientId)
+    {
+        NetworkObject.SpawnAsPlayerObject(clientId);
+        SetTeamTypeRpc(team);
+
+        UIInitRpc(team);
+    }
+
+    [Rpc(SendTo.ClientsAndHost)]
+    private void UIInitRpc(UnitTeamType team)
+    {
+        UIManager.Instance.Init(_hpController, team);
+        HeroHpBarUI heroHpBarUI = _unitHPBarUI as HeroHpBarUI;
 
         heroHpBarUI.UpdateName(GetHeroName());
         heroHpBarUI.UpdateLevel(GetLevel());
 
-        PlayerInputManager.Instance.OnLeftClickEvent.AddListener(OnLeftMouseDown);
-        PlayerInputManager.Instance.OnRightClickEvent.AddListener(OnRightMouseDown);
-        PlayerInputManager.Instance.OnAttackButtonEvent.AddListener(OnAttackButtonDown);
-
-        #region TEST
-        if (IsLocalPlayer)
-        {
-            gameObject.name = "PLAYER";
-            if (IsHost)
-            {
-                SetTeamTypeRpc(UnitTeamType.RedTeam);
-                UIManager.Instance.Init(_hpController, UnitTeamType.RedTeam);
-
-                heroHpBarUI.UpdateName("RED");
-            }
-            else
-            {
-                SetTeamTypeRpc(UnitTeamType.BlueTeam);
-                UIManager.Instance.Init(_hpController, UnitTeamType.BlueTeam);
-
-                heroHpBarUI.UpdateName("BLUE");
-            }
-        }
-        else
-        {
-            gameObject.name = "OTHER";
-            if (!IsHost)
-            {
-                UIManager.Instance.Init(_hpController, UnitTeamType.RedTeam);
-                heroHpBarUI.UpdateName("RED");
-            }
-            else
-            {
-                UIManager.Instance.Init(_hpController, UnitTeamType.BlueTeam);
-                heroHpBarUI.UpdateName("BLUE");
-            }
-        }
-
-        _hpController.Init(_unitStatusController.GetMaxHP());
         if (IsOwner)
         {
             FindAnyObjectByType<CinemachineCamera>().Follow = transform;
         }
-        #endregion
-    }
-
-    private void OnLeftMouseDown()
-    {
-        IsAttackButtonDown = false;
-    }
-
-    private void OnRightMouseDown()
-    {
-        IsAttackButtonDown = false;
     }
 
     private void OnAttackButtonDown()
@@ -96,7 +65,7 @@ public class PlayerController : UnitController
     public override void Dead()
     {
         IsDead = true;
-        _playerAction.StopMove();
+        StopMove();
         _collider.enabled = false;
 
         StartCoroutine(WaitRespawnCoroutine(RESPAWN_TIME[_unitStatusController.GetLevel()]));
@@ -136,13 +105,131 @@ public class PlayerController : UnitController
         _hpController.ChangeHPRpc(-damage);
     }
 
-    public override void OnTriggerEnterFromChild(Collider other)
+    private void Update()
     {
-        throw new NotImplementedException();
+        if (IsHost)
+        {
+            SetAnimatorBoolRpc("IsRun", _navMeshAgent.remainingDistance > _navMeshAgent.stoppingDistance);
+        }
+
+        if (!IsOwner)
+        {
+            return;
+        }
+
+        if (_target == null)
+        {
+            return;
+        }
+
+        if (_target.IsDead)
+        {
+            _target = null;
+            return;
+        }
+
+        if (IsTargetInAttackRange())
+        {
+            if (_isAttacking)
+            {
+                return;
+            }
+            else
+            {
+                _attackCoroutine = StartCoroutine(AttackCoroutine(_target, 1.0f, 1.0f));
+                StopMove();
+            }
+        }
+        else
+        {
+            float distanceToTarget = Vector3.Distance(transform.position, _target.transform.position);
+            if (distanceToTarget > MOVE_STOPPING_DISTANCE)
+            {
+                SetMoveDestinationRpc(_target.transform.position);
+            }
+            else
+            {
+                StopMove();
+            }
+        }
     }
 
-    public override void OnTriggerExitFromChild(Collider other)
+    [Rpc(SendTo.Server)]
+    private void SetAnimatorBoolRpc(string name, bool param)
     {
-        throw new NotImplementedException();
+        _modelAnimator.SetBool(name, param);
+    }
+
+    public void OnLeftMouseDown()
+    {
+        if (!IsOwner)
+        {
+            return;
+        }
+
+        IsAttackButtonDown = false;
+
+        if (IsDead)
+        {
+            Logger.Info("IsDead 켜져있음");
+            return;
+        }
+
+        if (IsAttackButtonDown)
+        {
+            Ray ray = Camera.main.ScreenPointToRay(Input.mousePosition);
+
+            if (Physics.Raycast(ray, out RaycastHit unitHit, 100f, _unitLayerMask))
+            {
+                GameObject hitObject = unitHit.collider.gameObject;
+
+                if (hitObject.TryGetComponent<UnitController>(out var unit))
+                {
+                    if (unit.TeamType != TeamType)
+                    {
+                        _target = unit;
+                    }
+                }
+            }
+            else if (Physics.Raycast(ray, out RaycastHit hit, 100f, _groundMask))
+            {
+                SetMoveDestinationRpc(hit.point);
+            }
+        }
+    }
+
+    public void OnRightMouseDown()
+    {
+        if (!IsOwner)
+        {
+            return;
+        }
+
+        IsAttackButtonDown = false;
+
+        if (IsDead)
+        {
+            Logger.Info("Is Dead");
+            return;
+        }
+
+        _target = null;
+
+        Ray ray = Camera.main.ScreenPointToRay(Input.mousePosition);
+
+        if (Physics.Raycast(ray, out RaycastHit unitHit, 100f, _unitLayerMask))
+        {
+            GameObject hitObject = unitHit.collider.gameObject;
+
+            if (hitObject.TryGetComponent<UnitController>(out var unit))
+            {
+                Logger.Info($"Mouse Hit Unit: {unit.name}");
+                _target = unit;
+            }
+        }
+        else if (Physics.Raycast(ray, out RaycastHit groundHit, 100f, _groundMask))
+        {
+            SetMoveDestinationRpc(groundHit.point);
+        }
     }
 }
