--- Assets/Scripts/Unit/UnitRangeTrigger.cs
@@ -1,16 +1,0 @@
-using UnityEngine;
-
-public class UnitRangeTrigger : MonoBehaviour
-{
-    [SerializeField] private UnitController _ownerUnit;
-
-    private void OnTriggerEnter(Collider other)
-    {
-        _ownerUnit.OnTriggerEnterFromChild(other);
-    }
-
-    private void OnTriggerExit(Collider other)
-    {
-        _ownerUnit.OnTriggerExitFromChild(other);
-    }
-}
