--- Assets/Scripts/Unit/UnitRangeTrigger.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: dac6feb15ee672e4683a911e344e4918
